로그인정보 받기 완

+++ src/main/java/com/example/tripit/user/dto/CustomUserDetails.java
@@ -71,7 +71,7 @@ public class CustomUserDetails  implements UserDetails {
         return true;
     }
 
-    public Long getUserId() {
-        return (long) userEntity.getUserId();
+    public Integer getUserId() {
+        return  userEntity.getUserId();
     }
 }
